memcache class added

+++ app/libraries/Cache/File.php
@@ -3,25 +3,26 @@
 final class Cache_File {
 
     private $_cachedir = '';
+    private $_validdir = false;
     private $_lkeydata = array();
 
     public function __construct() {
         $cachedir = APP_DIR . 'logs/';
-        if (!is_dir($cachedir) or ! is_writable($cachedir)) {
-            return;
+        if (is_dir($cachedir) && is_writable($cachedir)) {
+            $this->_validdir = true;
         }
         $this->_cachedir = $cachedir;
     }
 
     public function set($key, $data, $ttl = 180) {
-        $data = serialize(array(time() + $ttl, $data));
-        $file = $this->_cachedir . sha1($key);
-        if (file_exists($file)) {
-            unlink($file);
-        }
+        if ($this->_validdir) {
+            $data = serialize(array(time() + $ttl, $data));
+            $file = $this->_cachedir . sha1($key);
+            if (file_exists($file)) {
+                unlink($file);
+            }
 
-        if (!@file_put_contents($file, $data)) {
-            return;
+            @file_put_contents($file, $data);
         }
     }
 
@@ -30,19 +31,21 @@ final class Cache_File {
     }
 
     public function valid($key) {
-        $file1 = glob($this->_cachedir . sha1($key));
-        array_shift($file1);
-        $filename = $this->_cachedir . sha1($key);
-        if (file_exists($filename)) {
-            $data1 = @file_get_contents($filename);
-            $data = unserialize($data1);
+        if ($this->_validdir) {
+            $file1 = glob($this->_cachedir . sha1($key));
+            array_shift($file1);
+            $filename = $this->_cachedir . sha1($key);
+            if (file_exists($filename)) {
+                $data1 = @file_get_contents($filename);
+                $data = unserialize($data1);
 
-            if (time() > $data[0]) {
-                unlink($filename);
-                return false;
-            } else {
-                $this->_lkeydata[sha1($key)] = $data[1];
-                return true;
+                if (time() > $data[0]) {
+                    unlink($filename);
+                    return false;
+                } else {
+                    $this->_lkeydata[sha1($key)] = $data[1];
+                    return true;
+                }
             }
         }
         return false;
